Added a method to create a new department for the supervisor

+++ bamazon.sql
@@ -4,6 +4,7 @@ CREATE DATABASE bamazon;
 
 USE  bamazon;
 
+-- Creates a products table
 CREATE TABLE products(
     item_id INT NOT NULL AUTO_INCREMENT,
     product_name VARCHAR(50) NOT NULL,
@@ -12,6 +13,7 @@ CREATE TABLE products(
 stock_quantity INT NOT NULL,
 primary key(item_id));
 
+-- Inserts the data into products table
 INSERT INTO products(product_name,department_name,price,stock_quantity) values("Pen","Stationary",23.50,10);
 INSERT INTO products(product_name,department_name,price,stock_quantity) values("Eraser","Stationary",10,15);
 INSERT INTO products(product_name,department_name,price,stock_quantity) values("Water Colors","Stationary",50,20);
@@ -48,7 +50,7 @@ delimiter ;
  DELIMITER ;
 
 
--- This Sp displays the items with stock quantity less than 5
+-- This SP displays the items with stock quantity less than 5
  DELIMITER //
 Create procedure displayLowStock()
 begin
@@ -56,19 +58,23 @@ select item_id as ID,product_name as Product,stock_quantity as Quantity from bam
 end //
 delimiter ;
 
+-- Creats a department table
 CREATE TABLE departments(
     department_id INT NOT NULL AUTO_INCREMENT,
     department_name VARCHAR(50),
-    over_head_costs INT,
+    over_head_costs float,
     primary key (department_id)
 );
 
+-- This adds the new column to the existing table
 alter table products add column product_sales float not null ;
 
 insert into departments(department_name,over_head_costs) values("Stationary",30);
 insert into departments(department_name,over_head_costs) values("Footwear",35);
 insert into departments(department_name,over_head_costs) values("Electronics",80);
 
+
+-- This SP displays the department details and calculates the profits
 delimiter //
 create procedure displayDepartments()
 begin
